[core] ci_entrypoint: refactor run_ci to make it more general
