+++ bin/run_cli
@@ -1,0 +1,1 @@
+../projects/core/ci_entrypoint/run_cli.py
