+++ projects/core/ci_entrypoint/run_ci.py
@@ -17,555 +17,10 @@ Examples:
     run skeleton validate
 """
 
-# Import CI preparation module
-import logging
-import os
-import signal
-import subprocess
-import sys
-import time
-from pathlib import Path
+from projects.core.ci_entrypoint.run_common import main_orchestrator
 
-import click
-
-from projects.core.ci_entrypoint import prepare_ci
-
-
-def setup_logging():
-    """Set up logging configuration."""
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(levelname)s: %(message)s",
-        handlers=[logging.StreamHandler(sys.stderr)],
-    )
-
-
-# Set up logging
-
-logger = logging.getLogger(__name__)
-
-FORGE_HOME = Path(__file__).resolve().parent.parent.parent.parent
-
-EXTRA_PACKAGES = []
-
-# Global reference to child process for signal forwarding
-_child_process = None
-
-
-def signal_handler_sigint(sig, frame):
-    """Handle SIGINT (Ctrl+C) gracefully."""
-    logger.info("🚫 Received SIGINT (Ctrl+C) - Interrupting CI operation...")
-
-    # Forward signal to child process first
-    if _child_process and _child_process.poll() is None:  # Child is still running
-        logger.info("📡 Forwarding SIGINT to child process...")
-        try:
-            _child_process.send_signal(signal.SIGINT)
-        except (OSError, ProcessLookupError):
-            pass  # Child may have already terminated
-
-    # Emergency cleanup of dual output
-    prepare_ci.shutdown_dual_output()
-
-    sys.exit(130)  # Standard exit code for SIGINT
-
-
-def signal_handler_sigterm(sig, frame):
-    """Handle SIGTERM gracefully."""
-    logger.info("🛑 Received SIGTERM - Terminating CI operation...")
-
-    # Forward signal to child process first
-    if _child_process and _child_process.poll() is None:  # Child is still running
-        logger.info("📡 Forwarding SIGTERM to child process...")
-        try:
-            _child_process.send_signal(signal.SIGTERM)
-        except (OSError, ProcessLookupError):
-            pass  # Child may have already terminated
-
-    # Emergency cleanup of dual output
-    prepare_ci.shutdown_dual_output()
-
-    sys.exit(143)  # Standard exit code for SIGTERM
-
-
-def setup_signal_handlers():
-    """Set up signal handlers for graceful interruption."""
-    try:
-        signal.signal(signal.SIGINT, signal_handler_sigint)
-        signal.signal(signal.SIGTERM, signal_handler_sigterm)
-        # SIGPIPE handling for broken pipes
-        if hasattr(signal, "SIGPIPE"):
-            signal.signal(signal.SIGPIPE, signal.SIG_DFL)
-    except Exception:
-        # Signal handling might not be available on all platforms
-        pass
-
-
-def install_extra_packages(packages):
-    if not packages:
-        return
-
-    print(f"📦 Installing {'/'.join(packages)} packages...")
-
-    # Try uv first with no-cache to avoid permission issues
-    try:
-        subprocess.run(
-            ["uv", "pip", "install", "--no-cache", *packages],
-            check=True,
-            capture_output=True,
-        )
-        print(f"✅ {'/'.join(packages)} packages installed successfully with uv")
-        print()
-    except (FileNotFoundError, subprocess.CalledProcessError):
-        # Fallback to pip with user installation
-        try:
-            subprocess.run(
-                [
-                    sys.executable,
-                    "-m",
-                    "pip",
-                    "install",
-                    "--user",
-                    "--no-cache-dir",
-                    *packages,
-                ],
-                check=True,
-                capture_output=True,
-            )
-            print(f"✅ {'/'.join(packages)} packages installed successfully with pip")
-            print()
-        except subprocess.CalledProcessError as pip_error:
-            print(f"❌ Failed to install {'/'.join(packages)}: {pip_error}")
-            raise RuntimeError("failed to install the extra packages") from pip_error
-
-    # Ensure user site-packages is in path
-    import site
-
-    user_site = site.getusersitepackages()
-    if user_site not in sys.path:
-        sys.path.insert(0, user_site)
-
-    # Also check for common install locations
-    import os
-
-    possible_paths = [
-        os.path.expanduser("~/.local/lib/python3.11/site-packages"),
-        os.path.expanduser("~/.local/lib/python3.12/site-packages"),
-        os.path.expanduser("~/.local/lib/python3.13/site-packages"),
-        user_site,
-    ]
-
-    for path in possible_paths:
-        if os.path.exists(path) and path not in sys.path:
-            sys.path.insert(0, path)
-
-    # Clear import cache and try again
-    import importlib
-
-    importlib.invalidate_caches()
-
-
-def prepare():
-    setup_logging()
-
-    # Install click package using uv (as non-root user)
-    install_extra_packages(EXTRA_PACKAGES)
-
-    # Set up ARTIFACT_DIR
-    prepare_ci.precheck_artifact_dir()
-
-    if os.isatty(sys.stdin.fileno()):
-        # not working very well from a TTY ...
-        logger.info("Running from a TTY, not enabling the dual output")
-    elif (Path(os.environ["ARTIFACT_DIR"]) / "run.log").exists():
-        logger.info("run.log file already exists, not enabling the dual output")
-    else:
-        # Set up dual output as early as possible
-        prepare_ci.setup_dual_output()
-
-
-def find_project_directory(project_name: str) -> Path | None:
-    """
-    Find the directory for the specified project.
-
-    Args:
-        project_name: Name of the project to find
-
-    Returns:
-        Path to project directory if found, None otherwise
-    """
-    # Look in the projects directory
-    projects_dir = FORGE_HOME / "projects"
-    project_dir = projects_dir / project_name
-
-    if project_dir.exists() and project_dir.is_dir():
-        return project_dir
-
-    return None
-
-
-def find_ci_script(project_dir: Path, operation: str) -> Path | None:
-    """
-    Find the appropriate CI script for the operation.
-
-    Args:
-        project_dir: Project directory path
-        operation: Operation to perform (e.g., 'ci')
-
-    Returns:
-        Path to CI script if found, None otherwise
-    """
-    # Check possible locations for CI scripts
-    possible_locations = [
-        # Operation-specific script in orchestration subdirectory
-        project_dir / "orchestration" / f"{operation}.py",
-    ]
-
-    for script_path in possible_locations:
-        if script_path.exists() and os.access(script_path, os.X_OK):
-            return script_path
-
-    return None
-
-
-def get_available_projects() -> list[str]:
-    """Get list of available projects."""
-
-    projects_dir = FORGE_HOME / "projects"
-
-    if not projects_dir.exists():
-        return []
-
-    projects = []
-    for proj_dir in projects_dir.iterdir():
-        if not proj_dir.is_dir():
-            continue
-        if proj_dir.name.startswith("."):
-            continue
-
-        # Only include projects that have an orchestration directory
-        orchestration_dir = proj_dir / "orchestration"
-        if orchestration_dir.exists() and orchestration_dir.is_dir():
-            projects.append(proj_dir.name)
-
-    return sorted(projects)
-
-
-def list_projects():
-    """List all available projects."""
-    projects = get_available_projects()
-
-    if not projects:
-        click.echo("📂 No projects found")
-        return
-
-    click.echo("📂 Available projects:")
-    for project in projects:
-        project_dir = find_project_directory(project)
-        ci_script = find_ci_script(project_dir, "ci")
-        status = "✅" if ci_script else "⚠️"
-
-        click.echo(f"   {status} {project}")
-
-    click.echo()
-    click.echo("Usage:")
-    click.echo("   run <project> <operation>  # Execute project operation")
-    click.echo("   run projects               # List projects explicitly")
-    click.echo()
-    click.echo("Examples:")
-    click.echo(f"   run {projects[0]} prepare")
-    click.echo(f"   run {projects[0]} test")
-    if len(projects) > 1:
-        click.echo(f"   run {projects[1]} validate")
-
-
-def show_project_operations(project: str):
-    """Show available operations for a project by listing Python files."""
-    click.echo(f"🔧 Available operations for project '{project}':")
-
-    # Find project directory
-    project_dir = find_project_directory(project)
-    if not project_dir:
-        click.echo(click.style(f"❌ ERROR: Project '{project}' not found.", fg="red"), err=True)
-        return
-
-    python_files = []
-
-    def add_python_file(file_path):
-        if not file_path.is_file():
-            return
-        # Skip files that are not executable
-        if not os.access(file_path, os.X_OK):
-            return
-
-        operation_name = file_path.stem  # filename without .py extension
-        python_files.append((operation_name, file_path))
-
-    # List Python files in the orchestration directory
-    for file_path in (project_dir / "orchestration").glob("*.py"):
-        add_python_file(file_path)
-
-    if not python_files:
-        click.echo("⚠️  No Python files found in project directory")
-        click.echo(f"📁 Project directory: {project_dir}")
-        return
-
-    click.echo()
-    click.echo("📄 Available Python files:")
-
-    operation_files = []
-
-    for operation_name, file_path in sorted(python_files):
-        operation_files.append((operation_name, file_path))
-
-    for operation_name, _file_path in operation_files:
-        click.echo(f"   📝 {operation_name}.py")
-
-    click.echo(f"Usage: run {project} <filename_without_py>")
-    click.echo()
-    click.echo("Examples:")
-    for operation_name, _ in python_files[:3]:
-        click.echo(f"   run {project} {operation_name}")
-
-
-def parse_cli_help(help_output: str) -> list[str]:
-    """Parse CLI help output to extract available commands."""
-    operations = []
-    in_commands_section = False
-
-    lines = help_output.split("\n")
-    for line in lines:
-        line = line.strip()
-
-        # Look for "Commands:" section
-        if line.lower().startswith("commands:"):
-            in_commands_section = True
-            continue
-
-        # Stop when we hit another section
-        if in_commands_section and line and not line.startswith(" "):
-            break
-
-        # Extract command names
-        if in_commands_section and line.startswith(" "):
-            # Format is typically: "  command_name  Description"
-            parts = line.split()
-            if parts:
-                command_name = parts[0].strip()
-                # Skip common help/utility commands
-                if command_name not in ["--help", "--version", "-h", "-v"]:
-                    operations.append(command_name)
-
-    return operations
-
-
-def execute_project_operation(
-    project: str,
-    operation: str,
-    args: tuple,
-    verbose: bool = False,
-    dry_run: bool = False,
-    do_prepare_ci: bool = True,
-):
-    """Execute a project operation."""
-
-    if not (isinstance(args, list) or isinstance(args, tuple)):
-        raise ValueError(f"Args={args} must be a list, not {args.__class__.__name__}")
-
-    if verbose:
-        click.echo("")
-        click.echo("🚀 FORGE CI Orchestration")
-        click.echo(f"Project: {project}")
-        click.echo(f"Operation: {operation}")
-        click.echo(f"Arguments: {' '.join(args)}")
-        click.echo("")
-
-    # Execute CI preparation tasks
-    if do_prepare_ci:
-        try:
-            prepare_ci.prepare(
-                verbose=verbose,
-                project=project,
-                operation=operation,
-                args=list(args),
-            )
-        except:
-            click.echo(click.style("❌ ERROR: CI preparation failed", fg="red"), err=True)
-            raise
-    else:
-        logger.warning("CI preparation not enabled, skipping preparation")
-
-    # Find project directory
-    project_dir = find_project_directory(project)
-    if not project_dir:
-        click.echo(click.style(f"❌ ERROR: Project '{project}' not found.", fg="red"), err=True)
-
-        available_projects = get_available_projects()
-        if available_projects:
-            click.echo("\n📂 Available projects:")
-            for proj in available_projects:
-                click.echo(f"   • {proj}")
-        else:
-            click.echo("📂 No projects found in projects/ directory")
-
-        sys.exit(1)
-
-    # Find CI script
-    ci_script = find_ci_script(project_dir, operation)
-    if not ci_script:
-        script_path = project_dir.relative_to(FORGE_HOME) / "orchestration" / f"{operation}.py"
-
-        if script_path.exists():
-            click.echo(
-                click.style(
-                    f"❌ ERROR: CI script exists but is not executable for project '{project}' operation '{operation}'.",
-                    fg="red",
-                ),
-                err=True,
-            )
-            click.echo(f"💡 Fix with: chmod +x {script_path}")
-        else:
-            click.echo(
-                click.style(
-                    f"❌ ERROR: No CI script '{script_path}' found for project '{project}' operation '{operation}'.",
-                    fg="red",
-                ),
-                err=True,
-            )
-            click.echo(f"🔍 Expected: {script_path}")
-        sys.exit(1)
-
-    # Prepare command - don't pass operation as it's just the script name
-
-    cmd = [sys.executable, str(ci_script)] + list(args)
-
-    if verbose or dry_run:
-        click.echo("\n🔧 Execution Details:")
-        click.echo(f"   Command: {' '.join(cmd)}")
-        click.echo(f"   Working Directory: {Path.cwd()}")
-        click.echo(f"   Script: {ci_script}")
-
-    if dry_run:
-        click.echo("\n🧪 DRY RUN: Would execute the above command")
-        click.echo("✨ Use --verbose to see execution details without --dry-run")
-        return
-
-    # Execute the command
-    click.echo("")
-    click.echo(f"▶️  Executing {project} {operation} {' '.join(args)} | {' '.join(cmd)}")
-    click.echo("")
-
-    try:
-        # Track start time for duration calculation
-        start_time = time.time()
-
-        # Set up signal handlers for graceful interruption (safe to call multiple times)
-        setup_signal_handlers()
-
-        # Use Popen to get access to process object for signal forwarding
-        global _child_process
-        _child_process = subprocess.Popen(
-            cmd,
-            stdin=None,  # Inherit stdin for pdb/debugging
-            stdout=None,  # Inherit stdout for pdb/debugging
-            stderr=None,  # Inherit stderr for pdb/debugging
-        )
-
-        # Wait for process to complete
-        result_code = _child_process.wait()
-
-        # Create result object similar to subprocess.run()
-        class Result:
-            def __init__(self, returncode):
-                self.returncode = returncode
-
-        result = Result(result_code)
-        click.echo()
-        click.echo(
-            f"▶️  Execution of {project} {operation} {' '.join(args)} returned {result.returncode}"
-        )
-        click.echo()
-
-        finish_reason = (
-            prepare_ci.FinishReason.SUCCESS
-            if result.returncode == 0
-            else prepare_ci.FinishReason.ERROR
-        )
-
-        success = finish_reason == prepare_ci.FinishReason.SUCCESS
-        # Post-execution checks and status reporting
-        status_message = prepare_ci.postchecks(
-            project, operation, start_time, finish_reason, list(args)
-        )
-        msg = click.style(status_message, fg="green" if success else "red")
-
-        click.echo()
-        click.echo(msg, err=not success)
-
-        # Properly shutdown dual output to flush all buffers and terminate daemon
-        prepare_ci.shutdown_dual_output()
-
-        sys.exit(result.returncode)
-
-    except Exception as e:
-        logger.exception(f"Unexpected exception {e.__class__.__name__}: {e}")
-
-        # Emergency cleanup of dual output to prevent hanging
-        try:
-            prepare_ci.shutdown_dual_output()
-        except Exception:
-            pass  # Don't let cleanup errors mask the original error
-
-        click.echo(
-            click.style("❌ ERROR: Unexpected error during execution", fg="red"),
-            err=True,
-        )
-        sys.exit(1)
-
-
-@click.command()
-@click.argument("project", required=False)
-@click.argument("operation", required=False)
-@click.argument("args", nargs=-1)
-@click.option("--verbose", "-v", is_flag=True, help="Enable verbose output", default=True)
-@click.option("--dry-run", is_flag=True, help="Show what would be executed without running it")
-def main(project, operation, args, verbose, dry_run):
-    """
-    FORGE CI Orchestration Entrypoint.
-
-    \b
-    Usage:
-        run                           # List available projects
-        run <project> <operation>     # Execute project operation
-        run projects                  # Explicit project listing
-
-    \b
-    Examples:
-        run llm-d prepare
-        run llm-d test
-        run skeleton validate
-    """
-
-    prepare()
-
-    # No arguments - list projects
-    if not project:
-        list_projects()
-        return
-
-    # Special case: explicit "projects" command
-    if project == "projects":
-        list_projects()
-        return
-
-    # Need operation for project execution - show available operations
-    if not operation:
-        show_project_operations(project)
-        sys.exit(1)
-
-    # Execute project operation
-    execute_project_operation(project, operation, args, verbose, dry_run)
-
+# Create main function for CI mode
+main = main_orchestrator(use_cli=False)
 
 if __name__ == "__main__":
     main()
